feat: Allow clicking on tags in Tags page
In the Home and Documents pages one can click on a particular tag that a
document has to search for all documents with that tag, but in the Tags
page that functionality is missing. This commit replaces the `<Tag>`
with a `<TagLink>`, so that one can click on it to initiate a search.