feat: Sort tags alphabetically
This commit modifies the Tag repository to return tags in ordered
fashion, so that they appear sorted to the user. This may not be the
correct place to do so, and instead we might want to sort on the client
side, but that's a decision I cannot make.

It also includes a note in the repository test, but no implementation in
case this change is undesired.