feat(client): added Dutch translation (#607)
* feat: Add Dutch translation

This commit adds a Dutch translation by providing a `nl.dictionary.ts`
dictionary file, and adding itself to the i18n constants.

* fix: Add missing type to translations constant

This commit adds the missing type to the translations constant. This
type is not present on the English dictionary, but is present in all
others, and thus should also be present for the Dutch translation.

* fix: Make the linter happy

This commit makes the linter happy so that the PR can land.

* Update Dutch translation in thick-panthers-wash

---------

Co-authored-by: Corentin Thomasset <corentin.thomasset74@gmail.com>