feat(tags): allow clicking on tags in tags page (#609)
* feat: Allow clicking on tags in Tags page

In the Home and Documents pages one can click on a particular tag that a
document has to search for all documents with that tag, but in the Tags
page that functionality is missing. This commit replaces the `<Tag>`
with a `<TagLink>`, so that one can click on it to initiate a search.

* Enable tag clicks on Tags page

Made the tags clickable in the tag list

---------

Co-authored-by: Corentin Thomasset <corentin.thomasset74@gmail.com>